requirements.txt
keras
tensorflow
flask
matplotlib
pillow
numpy
werkzeug